Use exit code 4 when file generation is skipped
If the script succeeds, but selects the currently used mirror, the
current version returns 1. This is the same exit code used for actual
errors, like network problems, invalid input etc. This makes it
impossible for callers to know whether an actual error occurred or not.

Instead we now return 4 in this case.